% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/killSlurm.R
\name{killSlurm}
\alias{killSlurm}
\title{Kill a SLURM job from R}
\usage{
killSlurm(jobID, verbose = TRUE)
}
\arguments{
\item{jobID}{SLURM job number.  May be passed as a numeric value or as a
string.}

\item{verbose}{Logical indicating whether message should be printed regarding the job cancelation.}
}
\value{
Nothing.  Only a message indicating whether or not the job was
canceled.
}
\description{
Kill a SLURM job from R using the system command \code{scancel jobID}
}
\details{
If the job is not present in the queue, then \code{scancel} command is not
issued.  This function is used for error control in \code{\link{piclapply}}--but it can
also be called from the R terminal.
}
\author{
Landon Sego
}
\seealso{
\code{\link{piclapply}}.
}
\keyword{misc}

