% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piclapply-package.R
\docType{package}
\name{piclapply-package}
\alias{piclapply-package}
\title{piclapply:  Easy parallel computing with R on PIC}
\description{
An extension of lapply() that uses PIC to process the list in parallel
}
\details{
\tabular{ll}{
Package: \tab piclapply\cr
Type: \tab Package\cr
Version: \tab 1.1.3\cr
Date: \tab 2017-10-19\cr
License: \tab PNNL\cr }
}
\examples{
help(package = piclapply)
}
\author{
Landon Sego

Maintainer: Landon Sego <Landon.Sego@pnnl.gov>
}
\keyword{package}

