% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showAvailableSystems.R
\name{showAvailableSystems}
\alias{showAvailableSystems}
\title{Display PIC systems for which the 'piclapply' package currently works}
\usage{
showAvailableSystems(details = FALSE)
}
\arguments{
\item{details}{A logical indicating whether details should be displayed}
}
\value{
System names with details if requested
}
\description{
Display PIC systems for which the 'piclapply' package currently works
}
\details{
Extracts information contained in the source code file \code{systemSpecifics.R}.
This information is also contained in \code{getOption("piclapplyParms")}.
}

